requirements.txt
django==2.0.6
django-compressor==2.2
django-material==1.2.5
django-multiselectfield==0.1.8
django-registration==2.4.1
django-rest-framework==0.1.0
django-sass-processor==0.7
factory-boy==2.11.1
libsass==0.14.5
model-mommy==1.6.0
Pillow==5.1.0
psycopg2-binary==2.7.4
sorl-thumbnail==12.4.1
